fix:errorCheckServiceリファクタ #142
rename:common to errorCheck
